refactor(editor): Move settings.store and roles.store into @n8n/stores behind re-export shims (no-changelog)
Relocates settings.store and roles.store (with their colocated tests) from
editor-ui into the shared @n8n/stores package, exposed via subpath exports
(@n8n/stores/settings.store, @n8n/stores/roles.store). The previous
@/app/stores paths become deprecated re-export shims so existing importers
keep working and can be migrated per-directory later.

- Adds @n8n/rest-api-client and bowser as @n8n/stores dependencies.
- Replaces the editor-ui UserManagementAuthenticationMethod const enum with
  the runtime AuthenticationMethod value now exported from @n8n/api-types.
- Folds the INSECURE_CONNECTION_WARNING constant into the store.

Pure relocation — no behavior change.
