+++ packages/frontend/@n8n/stores/src/settings.store.test.ts
@@ -1,6 +1,7 @@
 import type { FrontendSettings } from '@n8n/api-types';
 import { createPinia, setActivePinia } from 'pinia';
 import { mock } from 'vitest-mock-extended';
+
 import { useSettingsStore } from './settings.store';
 
 const mockRootStore = {
@@ -45,14 +46,8 @@ vi.mock('@n8n/rest-api-client/api/events', () => ({
 	sessionStarted,
 }));
 
-vi.mock('@n8n/stores/useRootStore', () => ({
+vi.mock('./useRootStore', () => ({
 	useRootStore,
-}));
-
-vi.mock('@/app/stores/versions.store', () => ({
-	useVersionsStore: vi.fn(() => ({
-		initialize: vi.fn(),
-	})),
 }));
 
 vi.mock('@vueuse/core', async () => {
