+++ packages/frontend/editor-ui/src/Interface.ts
@@ -391,13 +391,6 @@ export interface IShareWorkflowsPayload {
 	shareWithIds: string[];
 }
 
-export const enum UserManagementAuthenticationMethod {
-	Email = 'email',
-	Ldap = 'ldap',
-	Saml = 'saml',
-	Oidc = 'oidc',
-}
-
 export interface IPermissionGroup {
 	loginStatus?: ILogInStatus[];
 	role?: Role[];
