+++ packages/frontend/editor-ui/src/app/constants/notice.ts
@@ -19,19 +19,3 @@ export const HIRING_BANNER = `
 
 Love n8n? Help us build the future of automation! https://n8n.io/careers?utm_source=n8n_user&utm_medium=console_output
 `;
-
-export const INSECURE_CONNECTION_WARNING = `
-<body style="margin-top: 20px; font-family: 'Open Sans', sans-serif; text-align: center;">
-<h1 style="font-size: 40px">&#x1F6AB;</h1>
-<h2>Your n8n server is configured to use a secure cookie, <br/>however you are either visiting this via an insecure URL, or using Safari.
-</h2>
-<br/>
-<div style="font-size: 18px; max-width: 640px; text-align: left; margin: 10px auto">
-	To fix this, please consider the following options:
-	<ul>
-		<li>Setup TLS/HTTPS (<strong>recommended</strong>), or</li>
-		<li>If you are running this locally, and not using Safari, try using <a href="http://localhost:5678">localhost</a> instead</li>
-		<li>If you prefer to disable this security feature (<strong>not recommended</strong>), set the environment variable <code>N8N_SECURE_COOKIE</code> to <code>false</code></li>
-	</ul>
-</div>
-</body>`;
