+++ packages/frontend/editor-ui/src/features/core/auth/views/SettingsPersonalView.test.ts
@@ -6,11 +6,10 @@ import { useSettingsStore } from '@/app/stores/settings.store';
 import { useUsersStore } from '@/features/settings/users/users.store';
 import { createComponentRenderer } from '@/__tests__/render';
 import { setupServer } from '@/__tests__/server';
-import { ROLE } from '@n8n/api-types';
+import { AuthenticationMethod, ROLE } from '@n8n/api-types';
 import { useUIStore } from '@/app/stores/ui.store';
 import { useCloudPlanStore } from '@/app/stores/cloudPlan.store';
 import { useSSOStore } from '@/features/settings/sso/sso.store';
-import { UserManagementAuthenticationMethod } from '@/Interface';
 
 let pinia: ReturnType<typeof createPinia>;
 let settingsStore: ReturnType<typeof useSettingsStore>;
@@ -54,7 +53,7 @@ describe('SettingsPersonalView', () => {
 
 		await settingsStore.getSettings();
 		ssoStore.initialize({
-			authenticationMethod: UserManagementAuthenticationMethod.Email,
+			authenticationMethod: AuthenticationMethod.Email,
 			config: settingsStore.settings.sso,
 			features: {
 				saml: true,
