+++ packages/frontend/editor-ui/src/features/settings/sso/sso.test.ts
@@ -1,7 +1,6 @@
-import type { OidcConfigDto } from '@n8n/api-types';
+import type { AuthenticationMethod, OidcConfigDto } from '@n8n/api-types';
 import { createPinia, setActivePinia } from 'pinia';
 import { useSSOStore, SupportedProtocols } from '@/features/settings/sso/sso.store';
-import type { UserManagementAuthenticationMethod } from '@/Interface';
 import * as ssoApi from '@n8n/rest-api-client/api/sso';
 
 vi.mock('@n8n/rest-api-client/api/sso');
@@ -24,7 +23,7 @@ describe('SSO store', () => {
 		'should check SSO login button availability when authenticationMethod is %s and enterprise feature is %s and sso login is set to %s',
 		(authenticationMethod, saml, loginEnabled, expectation) => {
 			ssoStore.initialize({
-				authenticationMethod: authenticationMethod as UserManagementAuthenticationMethod,
+				authenticationMethod: authenticationMethod as AuthenticationMethod,
 				config: {
 					saml: {
 						loginEnabled,
@@ -45,7 +44,7 @@ describe('SSO store', () => {
 		it('should populate callbackUrl when re-initialized with authenticated settings', () => {
 			// Simulate public settings (before login) — no callbackUrl
 			ssoStore.initialize({
-				authenticationMethod: 'oidc' as UserManagementAuthenticationMethod,
+				authenticationMethod: 'oidc' as AuthenticationMethod,
 				config: {
 					oidc: { loginEnabled: false, loginUrl: 'http://localhost:5678/rest/sso/oidc/login' },
 				},
@@ -56,7 +55,7 @@ describe('SSO store', () => {
 
 			// Simulate authenticated settings (after login) — includes callbackUrl
 			ssoStore.initialize({
-				authenticationMethod: 'oidc' as UserManagementAuthenticationMethod,
+				authenticationMethod: 'oidc' as AuthenticationMethod,
 				config: {
 					oidc: {
 						loginEnabled: false,
@@ -80,7 +79,7 @@ describe('SSO store', () => {
 		it('should initialize selectedAuthProtocol to OIDC when default authentication is OIDC', () => {
 			// Initialize with OIDC as default authentication method
 			ssoStore.initialize({
-				authenticationMethod: 'oidc' as UserManagementAuthenticationMethod,
+				authenticationMethod: 'oidc' as AuthenticationMethod,
 				config: {
 					oidc: { loginEnabled: true },
 				},
@@ -104,7 +103,7 @@ describe('SSO store', () => {
 		it('should initialize selectedAuthProtocol to SAML when default authentication is SAML', () => {
 			// Initialize with SAML as default authentication method
 			ssoStore.initialize({
-				authenticationMethod: 'saml' as UserManagementAuthenticationMethod,
+				authenticationMethod: 'saml' as AuthenticationMethod,
 				config: {
 					saml: { loginEnabled: true },
 				},
@@ -128,7 +127,7 @@ describe('SSO store', () => {
 		it('should initialize selectedAuthProtocol to SAML when default authentication is email', () => {
 			// Initialize with email as default authentication method
 			ssoStore.initialize({
-				authenticationMethod: 'email' as UserManagementAuthenticationMethod,
+				authenticationMethod: 'email' as AuthenticationMethod,
 				config: {},
 				features: {
 					saml: true,
@@ -150,7 +149,7 @@ describe('SSO store', () => {
 		it('should not reinitialize selectedAuthProtocol if already set', () => {
 			// Initialize with SAML as default authentication method
 			ssoStore.initialize({
-				authenticationMethod: 'saml' as UserManagementAuthenticationMethod,
+				authenticationMethod: 'saml' as AuthenticationMethod,
 				config: {
 					saml: { loginEnabled: true },
 				},
@@ -212,7 +211,7 @@ describe('SSO store', () => {
 		it('should reset oidc.loginEnabled to false when server config has it disabled', async () => {
 			// Start with loginEnabled = true via initialize
 			ssoStore.initialize({
-				authenticationMethod: 'oidc' as UserManagementAuthenticationMethod,
+				authenticationMethod: 'oidc' as AuthenticationMethod,
 				config: { oidc: { loginEnabled: true } },
 				features: { saml: false, ldap: false, oidc: true },
 			});
